requirements.txt
attrs==23.2.0
bidict==0.22.1
blinker==1.7.0
click==8.1.7
dnspython==2.5.0
email-validator==2.1.0.post1
Flask==2.2.3
Flask-Cors==4.0.0
flask-mongoengine==1.0.0
Flask-SocketIO==5.3.6
flask-validate-json==0.0.1
Flask-WTF==1.2.1
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mongoengine==0.27.0
numpy==1.26.3
opencv-python==4.9.0.80
pymongo==4.6.1
python-dotenv==1.0.1
python-engineio==4.8.2
python-socketio==5.11.0
referencing==0.33.0
rpds-py==0.17.1
simple-websocket==1.0.0
Werkzeug==2.2.2
wsproto==1.2.0
WTForms==3.1.1
zipp==3.17.0
